third_party/blink/web_tests/fast/dom/HTMLElement/iscontenteditable-designmodeon-subframe.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<script>
description('document.designMode should not inherit from parent documents.')

function runTest()
{
    evalAndLog('document.designMode = "on"');

    subDocument = document.getElementById("subframe").contentDocument;
    shouldBe('subDocument.designMode', '"off"');
    shouldBe('subDocument.getElementById("div").contentEditable', '"inherit"');
    shouldBe('subDocument.getElementById("div").isContentEditable', 'false');

    document.getElementById("subframe").style.display= 'none';
}
</script>
<iframe id="subframe" onload="runTest()" srcdoc="<div id='div'></div>"></iframe>
</body>
</html>
