third_party/blink/web_tests/fast/dom/HTMLElement/iscontenteditable-designmodeon-subframe-expected.txt

document.designMode should not inherit from parent documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

document.designMode = "on"
PASS subDocument.designMode is "off"
PASS subDocument.getElementById("div").contentEditable is "inherit"
PASS subDocument.getElementById("div").isContentEditable is false
